requirements.txt
Flask==3.1.0
python-dotenv==0.21.1
requests==2.32.3